streamlit_app/requirements.txt
altair==5.5.0
asttokens==3.0.0
attrs==25.3.0
bcrypt==4.3.0
blinker==1.9.0
cachetools==6.2.0
captcha==0.7.1
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
choreographer==1.0.10
click==8.1.8
cryptography==45.0.7
decorator==5.2.1
et_xmlfile==2.0.0
executing==2.2.1
extra-streamlit-components==0.1.81
gitdb==4.0.12
GitPython==3.1.45
greenlet==3.2.4
idna==3.10
iniconfig==2.1.0
ipython
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jsonpickle==4.1.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kaleido==1.1.0
logistro==1.1.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
narwhals==2.4.0
networkx
numpy
openpyxl==3.1.5
orjson==3.11.3
packaging==25.0
pandas==2.3.2
parso==0.8.5
pexpect==4.9.0
pillow==11.3.0
plotly==6.3.0
pluggy==1.6.0
prompt_toolkit==3.0.52
protobuf==6.32.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==21.0.0
pycparser==2.23
pydeck==0.9.1
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
pytest-timeout==2.4.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
pytz==2025.2
pyvis==0.3.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
simplejson==3.20.1
six==1.17.0
smmap==5.0.2
SQLAlchemy==2.0.43
st-theme==1.2.3
stack-data==0.6.3
streamlit==1.49.1
streamlit-aggrid
streamlit-authenticator
streamlit-card
streamlit-cookies-manager
streamlit-elements
streamlit-navigation-bar
streamlit-option-menu
tenacity==9.1.2
toml==0.10.2
tornado==6.5.2
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
wcwidth==0.2.13
xlsxwriter==3.2.5
pymysql